# ai-service/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeIQ AI Engine API
  description: '

    # ForgeIQ Industrial AI & Multi-Agent Operations Engine


    ForgeIQ is an AI-powered manufacturing intelligence and commerce platform for
    sheet-metal fabrication, CNC machining, laser cutting, and bending.


    ---


    ### Core Architectural Features

    - **100% Autonomous Local Intelligence**: Operates fully offline with zero runtime
    dependency on external LLM providers or OpenAI API keys.

    - **Deterministic Manufacturing Calculators**: Exact physics, material weight,
    laser cutting cycle time, and bending tonnage calculations run in pure deterministic
    Python algorithms outside the LLM.

    - **Tenant Isolation**: Every query and vector retrieval is strictly isolated
    by `X-Org-ID`.

    - **Structured JSON Observability**: Implements 12-factor JSON streaming logs
    with distributed request correlation (`X-Request-ID`) and performance alerts (>1000ms).


    ---


    ### Authentication & Multi-Tenancy Headers

    The microservice enforces the following headers for inter-service and client requests:

    - **`X-Org-ID`** *(string, required)*: Tenant organization ID (e.g. `org-forge-default`).

    - **`X-Service-Key`** *(string, optional in dev)*: Internal microservice authentication
    key.

    - **`X-User-ID`** *(string, optional)*: Authenticated user identifier (e.g. `usr-operator-42`).

...
    provided, user identity is automatically extracted.

...


    ---


...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/chat/completions:
    post:
      tags:
      - Chat & Copilot
      summary: Multi-Agent Manufacturing Operations Copilot
      description: 'Evaluates manufacturing inquiries, machine telemetry questions,
        or quotation calculations.

        Routes intent to specialized autonomous agents:

        - **ProductionAgent**: Machine capacity, bottleneck resolution, shift schedules.

        - **InventoryAgent**: Sheet stock levels, material reservations, warehouse
        location.

        - **QuotationAgent**: Laser cutting cycle times, bending tonnage, raw material
        scrap.

        - **AnalyticsAgent**: COGS, margin variances, scrap metrics.


        Enforces strict tenant isolation via `X-Org-ID` and retrieves context from
        the company''s private RAG index.'
      operationId: chat_completions_api_v1_chat_completions_post
      parameters:
      - name: X-Org-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: org-forge-default
          title: X-Org-Id
      - name: X-User-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: user-default
          title: X-User-Id
      - name: X-User-Role
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: factory_admin
          title: X-User-Role
      - name: X-Customer-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      - name: X-Service-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Service-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatQueryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIEnvelope_CopilotResponse_'
        '400':
          description: Bad Request — Missing required tenant headers (e.g. X-Org-ID)
            or malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — Missing or invalid X-Service-Key or invalid
            Authorization Bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — Requested resource or entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error — Unhandled exception; client should
            fallback to rule-based operations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat/customer:
    post:
      tags:
      - Chat & Copilot
      summary: Customer Portal Buyer Copilot
      description: 'Dedicated assistant for external buyers and customers.

        Provides order tracking, quotation explanations, and delivery ETAs strictly
        constrained

        to the authenticated `customer_id`. Guarantees zero cross-tenant or factory
        internal data leakage.'
      operationId: customer_chat_api_v1_chat_customer_post
      parameters:
      - name: X-Org-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: org-forge-default
          title: X-Org-Id
      - name: X-User-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: user-default
          title: X-User-Id
      - name: X-User-Role
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: factory_admin
          title: X-User-Role
      - name: X-Customer-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      - name: X-Service-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Service-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerChatRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIEnvelope_CopilotResponse_'
        '400':
          description: Bad Request — Missing required tenant headers (e.g. X-Org-ID)
            or malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — Missing or invalid X-Service-Key or invalid
            Authorization Bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — Requested resource or entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error — Unhandled exception; client should
            fallback to rule-based operations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    APIEnvelope_CopilotResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Indicates whether the request completed successfully
          default: true
          examples:
          - true
        data:
          anyOf:
          - $ref: '#/components/schemas/CopilotResponse'
          - type: 'null'
          description: Typed payload data returned by the endpoint
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Error message if success is false
          examples:
          - null
        warning:
          anyOf:
          - type: string
          - type: 'null'
          title: Warning
          description: Optional system warnings or human-verification notices
          examples:
          - null
        is_mock:
          type: boolean
          title: Is Mock
          description: Flag indicating if the response was served from a mock fallback
          default: false
          examples:
          - false
        provider_used:
          type: string
          title: Provider Used
          description: Name of the manufacturing AI or tool provider
          default: local
          examples:
          - local
          - quotation_agent
        latency_ms:
          type: number
          title: Latency Ms
          description: Server-side execution latency in milliseconds
          default: 0.0
          examples:
          - 3.6
      type: object
      title: APIEnvelope[CopilotResponse]
...
    AgentEvidence:
      properties:
        metric_name:
          type: string
          title: Metric Name
        value:
          title: Value
        confidence:
          type: number
          title: Confidence
          default: 1.0
        source:
          type: string
          title: Source
      type: object
      required:
      - metric_name
      - value
      - source
      title: AgentEvidence
    ChatQueryRequest:
      properties:
        query:
          type: string
          title: Query
          description: Natural language manufacturing inquiry, order query, or technical
            question
          examples:
          - What is the cutting cycle time and scrap rate for 100 units of 3mm SS304
            brackets?
        context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Context
          description: Optional operational context, e.g. current machine_id or active
            order_id
          examples:
          - machine_id: mach-laser-01
            stage: cutting
        conversation_history:
          anyOf:
          - items:
              additionalProperties:
                type: string
              type: object
            type: array
          - type: 'null'
          title: Conversation History
          description: Recent conversation turns for contextual dialog resolution
          examples:
          - - content: Can we take on 500 brackets?
              role: user
            - content: Checking laser capacity...
              role: assistant
      type: object
      required:
      - query
      title: ChatQueryRequest
    CopilotResponse:
      properties:
        answer:
          type: string
          title: Answer
        agent_routed:
          type: string
          title: Agent Routed
        supporting_evidence:
          items:
            $ref: '#/components/schemas/AgentEvidence'
          type: array
          title: Supporting Evidence
        citations:
          items:
            $ref: '#/components/schemas/RAGCitation'
          type: array
          title: Citations
        confidence:
          type: number
          title: Confidence
          default: 0.95
        recommendation:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommendation
        suggested_action:
          anyOf:
          - type: string
          - type: 'null'
          title: Suggested Action
        requires_approval:
          type: boolean
          title: Requires Approval
          default: false
        provider_used:
          type: string
          title: Provider Used
          default: mock
        is_mock:
          type: boolean
          title: Is Mock
          default: false
        latency_ms:
          type: number
          title: Latency Ms
          default: 0.0
      type: object
      required:
      - answer
      - agent_routed
      title: CopilotResponse
    CustomerChatRequest:
      properties:
        query:
          type: string
          title: Query
          description: Customer inquiry regarding order status, delivery date, or
            quotations
          examples:
          - What is the current production milestone for our order PO-2026-0492?
        customer_id:
          type: string
          title: Customer Id
          description: Authenticated customer account identifier for tenant data isolation
          examples:
          - cust-nexa-442
        order_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Order Id
          description: Optional specific purchase order or RFQ reference number
          examples:
          - PO-2026-0492
      type: object
      required:
      - query
      - customer_id
      title: CustomerChatRequest
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: Machine-readable error classification code
          examples:
          - INVALID_REQUEST
          - UNAUTHORIZED
          - NOT_FOUND
          - INTERNAL_SERVER_ERROR
        message:
          type: string
          title: Message
          description: Human-readable description of the error
          examples:
          - X-Org-ID header is strictly mandatory for tenant isolation
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Unique correlation ID for tracing the failed request in logs
          examples:
          - f07e6d61-970d-4733-8b98-cb20f67bab07
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        success:
          type: boolean
          title: Success
          description: Always false for error responses
          default: false
          examples:
          - false
        error:
          $ref: '#/components/schemas/ErrorDetail'
          description: Structured error payload
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional debugging context available in development environment
          examples:
          - null
      type: object
      required:
      - error
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RAGCitation:
      properties:
        source_id:
          type: string
          title: Source Id
        source_title:
          type: string
          title: Source Title
        source_type:
          type: string
          title: Source Type
        relevance_score:
          type: number
          title: Relevance Score
        snippet:
          type: string
          title: Snippet
        org_id:
          type: string
          title: Org Id
      type: object
      required:
      - source_id
      - source_title
      - source_type
      - relevance_score
      - snippet
      - org_id
      title: RAGCitation
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...